[#493] Sponsor 모델을 별도 파일로 분리
